:recycle: test(blog): consolidate auth and proxy tests under __tests__ layout
Bun mock.module() is process-wide. The pre-existing flat test files
(src/lib/auth.test.ts, src/app/api/proxy/route.test.ts) and the new
__tests__/-scoped files each registered competing mock.module() factories
for better-auth / @/lib/auth / next/server, so whichever factory was
registered last won and the other file's assertions broke.

Merge each module's tests into a single __tests__ file with one factory
per module and delete the flat legacy files:

  src/lib/__tests__/auth.test.ts
    - #496 betterAuth(options) introspection (6 cases)
    - requireSessionForRoute / requireSessionForPage (5 cases)

  src/app/api/proxy/__tests__/route.test.ts
    - Pre-existing: 401 unauthenticated, Content-Length header cap
    - SSRF hardening (#540): public 200, manual redirect rejection,
      MAX_REDIRECTS, DNS private-IP, DNS-rebind pin, timeout, stream cap

Pure test refactor: no behavior change, no source change.
